feat(Interaction): add interact object appearance script
The Interact Object Appearance script is used to determine the
appearance of a given object when an interactable object is interacted
with (touch/grab/use). It's possible to determine the state of the
object's renderer or even game object on any of the interaction states.

This script supersedes the functionality of the Interact Controller
Appearance script as this Interact Object Appearance script can be
used to hide the controller on interaction. This means the
Interact Controller Appearance script has been deprecated and all
of the example scenes that used it have been updated to use the new
Interact Object Appearance script.